+++ R/GamWeightData.R
@@ -1,0 +1,6 @@
+#' Gamma Weighted Data
+#'
+#' A dataset simulated containing training data from a Gamma Distribution
+#'
+#' @format A dataset with 600 rows and 4 variables with 1 label.
+"GamWeightData"