+++ NEWS.md
@@ -1,5 +1,12 @@
 # FuzzyClass (development version)
 
+# FuzzyClass 0.1.5
+
+- Add the "double weighted fuzzy gamma naive bayes classifier" as `DWFuzzyGammaNaiveBayes`
+- Add the "geometric fuzzy gamma naive bayes classifier" as `GeometricFuzzyGammaNaiveBayes`
+- Add the "Fuzzy Bayes Rule classifier" as `FuzzyBayesRule`
+- Remove the ordering method for functions with fuzzy parameters `metd=4`.
+
 # FuzzyClass 0.1.4
 
 - Added a new ordering method for functions with fuzzy parameters: `GauNBFuzzyParam`, `ExpNBFuzzyParam`,`GamNBFuzzyParam` and `PoiNBFuzzyParam`. The form can be used considering `metd = 4`.